Sketch out API call to dotcom
